Added folder creating & bug fixes.

+++ cmd/cli-file-manager/main.go
@@ -49,7 +49,7 @@ func initWidgets() {
 
 	p := widgets.NewParagraph()
 	p.Title = "Help Menu"
-	p.Text = "[↑](fg:green) - Scroll Up\n[↓](fg:green) - Scroll Down\n[q](fg:green) - Quit\n[Enter](fg:green) - Open\n[m](fg:green) - Memory Usage\n[f](fg:green) - Disk Information\n[^D (2 times)](fg:green) - Remove file\n[^F](fg:green) - Create file"
+	p.Text = "[↑](fg:green) - Scroll Up\n[↓](fg:green) - Scroll Down\n[q](fg:green) - Quit\n[Enter](fg:green) - Open\n[m](fg:green) - Memory Usage\n[f](fg:green) - Disk Information\n[^D (2 times)](fg:green) - Remove file\n[^F](fg:green) - Create file\n[^N](fg:green) - Create folder"
 	p.SetRect(35, 0, 70, 15)
 	p.BorderStyle.Fg = ui.ColorBlue
 	p.TitleStyle.Modifier = ui.ModifierBold
@@ -81,6 +81,7 @@ func initWidgets() {
 	previousKey := ""
 	inputField := ""
 	fileCreatingInProgress := false
+	dirCreatingInProgress := false
 	uiEvents := ui.PollEvents()
 	for {
 		e := <-uiEvents
@@ -88,48 +89,55 @@ func initWidgets() {
 		case "q", "<C-c>":
 			return
 		case "<Down>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				l.ScrollDown()
 				p2.Text = cfm.GetFileInformations(fmt.Sprintf("%v/%v", path, getFileName(l.SelectedRow)))
 			}
 		case "<Up>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				l.ScrollUp()
 				p2.Text = cfm.GetFileInformations(fmt.Sprintf("%v/%v", path, getFileName(l.SelectedRow)))
 			}
 		case "<Home>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				l.ScrollTop()
 				p2.Text = cfm.GetFileInformations(fmt.Sprintf("%v/%v", path, getFileName(l.SelectedRow)))
 			}
 		case "<End>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				l.ScrollBottom()
 				p2.Text = cfm.GetFileInformations(fmt.Sprintf("%v/%v", path, getFileName(l.SelectedRow)))
 			}
 		case "<C-d>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				if previousKey == "<C-d>" {
 					selected := getFileName(l.SelectedRow)
-					filePath := ""
-					if path[len(path)-1] == '/' || selected[0] == '/' {
-						filePath = fmt.Sprintf("%v%v", path, selected)
-					} else {
-						filePath = fmt.Sprintf("%v/%v", path, selected)
+					if selected != ".." && selected != "../" {
+						filePath := ""
+						if path[len(path)-1] == '/' || selected[0] == '/' {
+							filePath = fmt.Sprintf("%v%v", path, selected)
+						} else {
+							filePath = fmt.Sprintf("%v/%v", path, selected)
+						}
+						err := os.Remove(filePath)
+						if err == nil {
+							l.Rows = cfm.ReadFiles(path)
+							l.SelectedRow = 0
+							p2.Text = cfm.GetFileInformations(fmt.Sprintf("%v/%v", path, getFileName(l.SelectedRow)))
+						}
 					}
-					err := os.Remove(filePath)
-					if err != nil {
-						log.Fatal(e)
-						return
-					}
-					l.Rows = cfm.ReadFiles(path)
-					l.SelectedRow = 0
 				}
 			}
 		case "<C-f>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				fileCreatingInProgress = true
 				l.Rows = append(l.Rows, fmt.Sprintf("[?] %v", inputField))
+				l.SelectedRow = len(l.Rows) - 1
+			}
+		case "<C-n>":
+			if !fileCreatingInProgress && !dirCreatingInProgress {
+				dirCreatingInProgress = true
+				l.Rows = append(l.Rows, fmt.Sprintf("[$] %v", inputField))
 				l.SelectedRow = len(l.Rows) - 1
 			}
 		case "<Escape>":
@@ -138,19 +146,24 @@ func initWidgets() {
 				inputField = ""
 				l.SelectedRow = 0
 				l.Rows = l.Rows[:len(l.Rows)-1]
+			} else if dirCreatingInProgress {
+				dirCreatingInProgress = false
+				inputField = ""
+				l.SelectedRow = 0
+				l.Rows = l.Rows[:len(l.Rows)-1]
 			}
 		case "m":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				p3.Title = "Memory Usage"
 				p3.Text = cfm.ReadMemStats()
 			}
 		case "f":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				p3.Title = "Disk Information"
 				p3.Text = fmt.Sprintf("[All: ](fg:green) - %.2f GB\n[Used:](fg:green) - %.2f GB\n[Free:](fg:green) - %.2f GB", float64(disk.All)/float64(1024*1024*1024), float64(disk.Used)/float64(1024*1024*1024), float64(disk.Free)/float64(1024*1024*1024))
 			}
 		case "<Enter>":
-			if !fileCreatingInProgress {
+			if !fileCreatingInProgress && !dirCreatingInProgress {
 				selected := getFileName(l.SelectedRow)
 				if selected[len(selected)-1] == '/' {
 					if selected == "../" {
@@ -191,6 +204,16 @@ func initWidgets() {
 						fileCreatingInProgress = false
 					}
 				}
+			} else if dirCreatingInProgress {
+				if len(inputField) >= 3 {
+					err := os.Mkdir(fmt.Sprintf("%v/%v", path, inputField), 0755)
+					if err == nil {
+						l.Rows = cfm.ReadFiles(path)
+						l.SelectedRow = 0
+						inputField = ""
+						dirCreatingInProgress = false
+					}
+				}
 			}
 		}
 
@@ -205,9 +228,20 @@ func initWidgets() {
 				}
 				l.Rows[len(l.Rows)-1] = fmt.Sprintf("[?] %v", inputField)
 			}
+		} else if dirCreatingInProgress {
+			if e.ID[0] != '<' {
+				inputField = inputField + e.ID
+				l.Rows[len(l.Rows)-1] = fmt.Sprintf("[$] %v", inputField)
+			} else if e.ID == "<Backspace>" {
+				le := len(inputField)
+				if le > 0 {
+					inputField = inputField[:le-1]
+				}
+				l.Rows[len(l.Rows)-1] = fmt.Sprintf("[$] %v", inputField)
+			}
 		}
 
-		if !fileCreatingInProgress {
+		if !fileCreatingInProgress && !dirCreatingInProgress {
 			if previousKey == "<C-d>" {
 				previousKey = ""
 			} else {
